remove functions related to load_to_rds
